Transform virtme/guest into a Python package
setup.py doesn't copy the necessary bash files from virtme/guest folder
to the installed folder.

This can be solved if we transform virtme/guest folder, which only have bash
scripts, into a python package by add a __init__.py file and instructing
setup.py that virtme.guest is a package and we will need its contents as
package data.

This idea was based on the setuptools docs advice:
https://setuptools.readthedocs.io/en/latest/setuptools.html#non-package-data-files